for named arguments to person(), recognize substrings
R allows you to shorten named argument references to the shortest
unique substring for the defined named arguments in the function. For
example, if person() is defined as person(given = NULL, role = NULL),
it can be called as person(r = "aut", gi = "Me") and the arguments
will be recognized as role and given, respectively.